Convert variable into factor and keep value labels

This function converts a variable into a factor, but preserves variable and value label attributes.

as_factor(x, ..., add.non.labelled = FALSE)

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples'.

add.non.labelled
Logical, if TRUE, non-labelled values also get value labels.

Value

A factor, including variable and value labels. If x is a data frame, the complete data frame x will be returned, where variables specified in ... are coerced to factors (including variable and value labels); if ... is not specified, applies to all variables in the data frame.

Details

as_factor converts numeric values into a factor with numeric levels. as_label, however, converts a vector into a factor and uses value labels as factor levels.

Note

This function is intended for use with vectors that have value and variable label attributes. Unlike as.factor, as_factor converts a variable into a factor and preserves the value and variable label attributes.

Adding label attributes is automatically done by importing data sets with one of the read_*-functions, like read_spss. Else, value and variable labels can be manually added to vectors with set_labels and set_label.

Examples

library(sjmisc)

#> 
#> Attaching package: ‘sjmisc’

#> The following object is masked from ‘package:sjlabelled’:
#> 
#>     is_float

data(efc)
# normal factor conversion, loses value attributes
x <- as.factor(efc$e42dep)
frq(x)

#> 
#> # x <categorical> 
#> # total N=908  valid N=901  mean=2.94  sd=0.94
#>  
#>   val frq raw.prc valid.prc cum.prc
#>     1  66    7.27      7.33    7.33
#>     2 225   24.78     24.97   32.30
#>     3 306   33.70     33.96   66.26
#>     4 304   33.48     33.74  100.00
#>  <NA>   7    0.77        NA      NA
#> 
#> 

# factor conversion, which keeps value attributes
x <- as_factor(efc$e42dep)

#> Error in UseMethod("as_factor"): no applicable method for 'as_factor' applied to an object of class "c('double', 'numeric')"

frq(x)

#> 
#> # x <categorical> 
#> # total N=908  valid N=901  mean=2.94  sd=0.94
#>  
#>   val frq raw.prc valid.prc cum.prc
#>     1  66    7.27      7.33    7.33
#>     2 225   24.78     24.97   32.30
#>     3 306   33.70     33.96   66.26
#>     4 304   33.48     33.74  100.00
#>  <NA>   7    0.77        NA      NA
#> 
#> 

# create parially labelled vector
x <- set_labels(
  efc$e42dep,
  labels = c(
    `1` = "independent",
    `4` = "severe dependency",
    `9` = "missing value"
 ))

# only copy existing value labels
as_factor(x)

#> Error in UseMethod("as_factor"): no applicable method for 'as_factor' applied to an object of class "c('double', 'numeric')"

get_labels(as_factor(x), include.values = "p")

#> Error in UseMethod("as_factor"): no applicable method for 'as_factor' applied to an object of class "c('double', 'numeric')"

# also add labels to non-labelled values
as_factor(x, add.non.labelled = TRUE)

#> Error in UseMethod("as_factor"): no applicable method for 'as_factor' applied to an object of class "c('double', 'numeric')"

get_labels(as_factor(x, add.non.labelled = TRUE), include.values = "p")

#> Error in UseMethod("as_factor"): no applicable method for 'as_factor' applied to an object of class "c('double', 'numeric')"

# easily coerce specific variables in a data frame to factor
# and keep other variables, with their class preserved
as_factor(efc, e42dep, e16sex, c172code)

#>     c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#> 1        16        2      2     83      3       3       2       2       2
#> 2       148        2      2     88      3       3       3       3       3
#> 3        70        1      2     82      3       2       2       1       4
#> 4       168        1      2     67      4       4       1       3       1
#> 5       168        2      2     84      4       3       2       1       2
#> 6        16        2      2     85      4       2       2       3       3
#> 7       161        1      1     74      4       4       2       4       1
#> 8       110        4      2     87      4       3       2       2       1
#> 9        28        2      2     79      4       3       2       3       2
#> 10       40        2      2     83      4       3       2       1       2
#> 11      100        1      1     68      4       3       4       4       4
#> 12       25        8      2     97      3       3       3       3       1
#> 13       25        2      2     80      4       3       2       2       2
#> 14       24        1      2     75      3       3       2       4       4
#> 15       56        2      2     82      3       2       3       3       3
#> 16       20        2      2     89      3       4       2       1       3
#> 17       25        1      1     80      1       3       2       1       2
#> 18      126        1      1     72      3       4       2       1       2
#> 19      168        2      1     94      3       3       2       1       2
#> 20      118        1      1     79      4       3       2       4       2
#> 21      150        2      2     89      4       3       2       3       2
#> 22       50        1      1     67      3       4       1       2       1
#> 23       18        2      2     94      4       3       1       1       1
#> 24      168        2      2     83      4       3       2       2       4
#> 25       15        2      2     85      3       3       2       1       2
#> 26      168        1      1     80      4       2       3       4       4
#> 27       12        2      2     88      4       2       3       3      NA
#> 28        7        1      1     76      2       2       2       2       2
#> 29       35        2      2     84      3       3       2       2       2
#> 30      168        2      2     95      4       2       2       1       2
#> 31      150        2      1     88      4       3       2       4       4
#> 32      168        1      1     87      4       3       1       3       1
#> 33      168        4      2     89      3       3       2       2       3
#> 34      119        1      1     80      4       2       2       4       4
#> 35      168        1      2     75      4       3       2       1       2
#> 36      168        1      1     82      4       3       2       1       2
#> 37      168        1      1     69      4       3       4       3       2
#> 38       28        2      2     91      2       3       1       1       2
#>     c86cop5 c87cop6 c88cop7 c89cop8 c90cop9 c160age c161sex c172code c175empl
#> 1         1       1       2       3       3      56       2        2        1
#> 2         4       1       3       2       2      54       2        2        1
#> 3         1       1       1       4       3      80       1        1        0
#> 4         1       1       1       2       4      69       1        2        0
#> 5         2       2       1       4       4      47       2        2        0
#> 6         3       2       2       1       1      56       1        2        1
#> 7         1       2       4       1       4      61       2        2        0
#> 8         1       1       2       3       3      67       2        2        0
#> 9         2       1       3       1       3      59       2       NA        0
#> 10        1       1       1       1       3      49       2        2        0
#> 11        4       4       4       1       1      66       2        2        0
#> 12        3       1       4       3       1      47       2        2        1
#> 13        2       1       2       4       4      58       2        3        0
#> 14        1       1       2       4       4      75       1        1        0
#> 15        2       2       1       1       1      49       2        3        1
#> 16        3       1       2       1       3      56       2        2        0
#> 17        1       1       2       4       4      75       2        2        0
#> 18        1       1       2       3       3      70       2        2        0
#> 19        2       1       3       1       4      52       1        3        1
#> 20        1       3       3       2       2      48       2        3        1
#> 21        1       1       1       1       3      58       2        2        1
#> 22        1       1       1       2       2      65       2        2        0
#> 23        1       1       1       2       4      49       2        2        1
#> 24        1       1       2       3       4      60       2        2        0
#> 25        1       1       2       3       3      55       2        2        0
#> 26        3       1       4       1       4      62       2        2        0
#> 27        1       1       4       2       2      68       2        3        0
#> 28        1       1       1       3       2      76       2        2        0
#> 29        4       1       2       2       1      58       2        2        0
#> 30        1       1       1       2       4      65       2        1        0
#> 31        4       4       4       1       3      63       1        2        1
#> 32        1       1       1       1       4      79       2        1        0
#> 33        1       1       3       1       1      65       2        2        0
#> 34        2       1       4       2       3      74       2        2        0
#> 35        1       1       1       3       4      76       1        2        0
#> 36        1       1       1       4       4      73       2        1        0
#> 37        1       1       4       1       1      67       2        2        0
#> 38        1       1       1       4       4      62       2        2        0
#>     barthtot neg_c_7 pos_v_4 quol_5 resttotn tot_sc_e n4pstu nur_pst
#> 1         75      12      12     14        0        4      0      NA
#> 2         75      20      11     10        4        0      0      NA
#> 3         35      11      13      7        0        1      2       2
#> 4          0      10      15     12        2        0      3       3
#> 5         25      12      15     19        2        1      2       2
#> 6         60      19       9      8        1        3      2       2
#> 7          5      15      13     20        0        0      3       3
#> 8         35      11      14     20        0        1      1       1
#> 9         15      15      13      8        0        2      3       3
#> 10         0      10      13     15        1        1      3       3
#> 11        25      28       9      1        1        1      3       3
#> 12        85      18       8     19        1        1      1       1
#> 13        15      13      14     12        0        3      3       3
#> 14        70      18      14      8        0        0      1       1
#> 15        NA      16       9      8        3        3      0      NA
#> 16         0      13      14      6        0        2      0      NA
#> 17        95      11      15     16        0        2      0      NA
#> 18        55      11      13     14        0        0      2       2
#> 19        55      13      13     15        3        1      1       1
#> 20        45      17      12     10        0        7      2       2
#> 21         0      11      11     15        1        1      4       3
#> 22        70       9      12     21        0        0      1       1
#> 23        20       8      15     18        1        0      2       2
#> 24         0      14      15     15        1        1      3       3
#> 25        95      11      14     16        0        2      1       1
#> 26        10      23      12      0        1        1      2       2
#> 27        75      NA      10     10        0        5      2       2
#> 28        90      11       9     14        0        2      0      NA
#> 29        60      15      11      4        0        2      2       2
#> 30        20      11      NA     15        0        0      3       3
#> 31        NA      25      13      9        1        2      3       3
#> 32        50       9      12     18        0        1      2       2
#> 33        75      15       9     12        0        1      1       1
#> 34        50      20      12      1        0        7      3       3
#> 35        30       9      15      4        0        4      3       3
#> 36         0      10      14     13        0        1      3       3
#> 37        30      19       9      4        0        6      3       3
#> 38        90       8      14     16        1        1      0      NA
#>  [ reached getOption("max.print") -- omitted 870 rows ]

# use select-helpers from dplyr-package
library(dplyr)
as_factor(efc, contains("cop"), c161sex:c175empl)

#>     c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#> 1        16        2      2     83      3       3       2       2       2
#> 2       148        2      2     88      3       3       3       3       3
#> 3        70        1      2     82      3       2       2       1       4
#> 4       168        1      2     67      4       4       1       3       1
#> 5       168        2      2     84      4       3       2       1       2
#> 6        16        2      2     85      4       2       2       3       3
#> 7       161        1      1     74      4       4       2       4       1
#> 8       110        4      2     87      4       3       2       2       1
#> 9        28        2      2     79      4       3       2       3       2
#> 10       40        2      2     83      4       3       2       1       2
#> 11      100        1      1     68      4       3       4       4       4
#> 12       25        8      2     97      3       3       3       3       1
#> 13       25        2      2     80      4       3       2       2       2
#> 14       24        1      2     75      3       3       2       4       4
#> 15       56        2      2     82      3       2       3       3       3
#> 16       20        2      2     89      3       4       2       1       3
#> 17       25        1      1     80      1       3       2       1       2
#> 18      126        1      1     72      3       4       2       1       2
#> 19      168        2      1     94      3       3       2       1       2
#> 20      118        1      1     79      4       3       2       4       2
#> 21      150        2      2     89      4       3       2       3       2
#> 22       50        1      1     67      3       4       1       2       1
#> 23       18        2      2     94      4       3       1       1       1
#> 24      168        2      2     83      4       3       2       2       4
#> 25       15        2      2     85      3       3       2       1       2
#> 26      168        1      1     80      4       2       3       4       4
#> 27       12        2      2     88      4       2       3       3      NA
#> 28        7        1      1     76      2       2       2       2       2
#> 29       35        2      2     84      3       3       2       2       2
#> 30      168        2      2     95      4       2       2       1       2
#> 31      150        2      1     88      4       3       2       4       4
#> 32      168        1      1     87      4       3       1       3       1
#> 33      168        4      2     89      3       3       2       2       3
#> 34      119        1      1     80      4       2       2       4       4
#> 35      168        1      2     75      4       3       2       1       2
#> 36      168        1      1     82      4       3       2       1       2
#> 37      168        1      1     69      4       3       4       3       2
#> 38       28        2      2     91      2       3       1       1       2
#>     c86cop5 c87cop6 c88cop7 c89cop8 c90cop9 c160age c161sex c172code c175empl
#> 1         1       1       2       3       3      56       2        2        1
#> 2         4       1       3       2       2      54       2        2        1
#> 3         1       1       1       4       3      80       1        1        0
#> 4         1       1       1       2       4      69       1        2        0
#> 5         2       2       1       4       4      47       2        2        0
#> 6         3       2       2       1       1      56       1        2        1
#> 7         1       2       4       1       4      61       2        2        0
#> 8         1       1       2       3       3      67       2        2        0
#> 9         2       1       3       1       3      59       2       NA        0
#> 10        1       1       1       1       3      49       2        2        0
#> 11        4       4       4       1       1      66       2        2        0
#> 12        3       1       4       3       1      47       2        2        1
#> 13        2       1       2       4       4      58       2        3        0
#> 14        1       1       2       4       4      75       1        1        0
#> 15        2       2       1       1       1      49       2        3        1
#> 16        3       1       2       1       3      56       2        2        0
#> 17        1       1       2       4       4      75       2        2        0
#> 18        1       1       2       3       3      70       2        2        0
#> 19        2       1       3       1       4      52       1        3        1
#> 20        1       3       3       2       2      48       2        3        1
#> 21        1       1       1       1       3      58       2        2        1
#> 22        1       1       1       2       2      65       2        2        0
#> 23        1       1       1       2       4      49       2        2        1
#> 24        1       1       2       3       4      60       2        2        0
#> 25        1       1       2       3       3      55       2        2        0
#> 26        3       1       4       1       4      62       2        2        0
#> 27        1       1       4       2       2      68       2        3        0
#> 28        1       1       1       3       2      76       2        2        0
#> 29        4       1       2       2       1      58       2        2        0
#> 30        1       1       1       2       4      65       2        1        0
#> 31        4       4       4       1       3      63       1        2        1
#> 32        1       1       1       1       4      79       2        1        0
#> 33        1       1       3       1       1      65       2        2        0
#> 34        2       1       4       2       3      74       2        2        0
#> 35        1       1       1       3       4      76       1        2        0
#> 36        1       1       1       4       4      73       2        1        0
#> 37        1       1       4       1       1      67       2        2        0
#> 38        1       1       1       4       4      62       2        2        0
#>     barthtot neg_c_7 pos_v_4 quol_5 resttotn tot_sc_e n4pstu nur_pst
#> 1         75      12      12     14        0        4      0      NA
#> 2         75      20      11     10        4        0      0      NA
#> 3         35      11      13      7        0        1      2       2
#> 4          0      10      15     12        2        0      3       3
#> 5         25      12      15     19        2        1      2       2
#> 6         60      19       9      8        1        3      2       2
#> 7          5      15      13     20        0        0      3       3
#> 8         35      11      14     20        0        1      1       1
#> 9         15      15      13      8        0        2      3       3
#> 10         0      10      13     15        1        1      3       3
#> 11        25      28       9      1        1        1      3       3
#> 12        85      18       8     19        1        1      1       1
#> 13        15      13      14     12        0        3      3       3
#> 14        70      18      14      8        0        0      1       1
#> 15        NA      16       9      8        3        3      0      NA
#> 16         0      13      14      6        0        2      0      NA
#> 17        95      11      15     16        0        2      0      NA
#> 18        55      11      13     14        0        0      2       2
#> 19        55      13      13     15        3        1      1       1
#> 20        45      17      12     10        0        7      2       2
#> 21         0      11      11     15        1        1      4       3
#> 22        70       9      12     21        0        0      1       1
#> 23        20       8      15     18        1        0      2       2
#> 24         0      14      15     15        1        1      3       3
#> 25        95      11      14     16        0        2      1       1
#> 26        10      23      12      0        1        1      2       2
#> 27        75      NA      10     10        0        5      2       2
#> 28        90      11       9     14        0        2      0      NA
#> 29        60      15      11      4        0        2      2       2
#> 30        20      11      NA     15        0        0      3       3
#> 31        NA      25      13      9        1        2      3       3
#> 32        50       9      12     18        0        1      2       2
#> 33        75      15       9     12        0        1      1       1
#> 34        50      20      12      1        0        7      3       3
#> 35        30       9      15      4        0        4      3       3
#> 36         0      10      14     13        0        1      3       3
#> 37        30      19       9      4        0        6      3       3
#> 38        90       8      14     16        1        1      0      NA
#>  [ reached getOption("max.print") -- omitted 870 rows ]