Convert variable into factor with associated value labels

as_label() converts (replaces) values of a variable (also of factors or character vectors) with their associated value labels. Might be helpful for factor variables. For instance, if you have a Gender variable with 0/1 value, and associated labels are male/female, this function would convert all 0 to male and all 1 to female and returns the new variable as factor. as_character() does the same as as_label(), but returns a character vector.

as_character(x, ...)

# S3 method for data.frame
as_character(x, ..., add.non.labelled = FALSE,
  prefix = FALSE, var.label = NULL, drop.na = TRUE,
  drop.levels = FALSE, keep.labels = FALSE)

as_label(x, ...)

# S3 method for data.frame
as_label(x, ..., add.non.labelled = FALSE,
  prefix = FALSE, var.label = NULL, drop.na = TRUE,
  drop.levels = FALSE, keep.labels = FALSE)

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples'.

add.non.labelled
Logical, if TRUE, values without associated value label will also be converted to labels (as is). See 'Examples'.

prefix
Logical, if TRUE, the value labels used as factor levels or character values will be prefixed with their associated values. See 'Examples'.

var.label
Optional string, to set variable label attribute for the returned variable (see vignette Labelled Data and the sjlabelled-Package). If NULL (default), variable label attribute of x will be used (if present). If empty, variable label attributes will be removed.

drop.na
Logical, if TRUE, tagged NA values with value labels will be converted to regular NA's. Else, tagged NA values will be replaced with their value labels. See 'Examples' and get_na.

drop.levels
Logical, if TRUE, unused factor levels will be dropped (i.e. droplevels will be applied before returning the result).

keep.labels
Logical, if TRUE, value labels are preserved This allows users to quickly convert back factors to numeric vectors with as_numeric().

Value

A factor with the associated value labels as factor levels. If x is a data frame, the complete data frame x will be returned, where variables specified in ... are coerced to factors; if ... is not specified, applies to all variables in the data frame. as_character() returns a character vector.

Details

See 'Details' in get_na.

Note

Value label attributes (see get_labels) will be removed when converting variables to factors.

Examples

data(efc)
print(get_labels(efc)['c161sex'])

#> $c161sex
#> [1] "Male"   "Female"
#> 

head(efc$c161sex)

#> [1] 2 2 1 1 2 1

head(as_label(efc$c161sex))

#> [1] Female Female Male   Male   Female Male  
#> Levels: Male Female

print(get_labels(efc)['e42dep'])

#> $e42dep
#> [1] "independent"          "slightly dependent"   "moderately dependent"
#> [4] "severely dependent"  
#> 

table(efc$e42dep)

#> 
#>   1   2   3   4 
#>  66 225 306 304 

table(as_label(efc$e42dep))

#> 
#>          independent   slightly dependent moderately dependent 
#>                   66                  225                  306 
#>   severely dependent 
#>                  304 

head(efc$e42dep)

#> [1] 3 3 3 4 4 4

head(as_label(efc$e42dep))

#> [1] moderately dependent moderately dependent moderately dependent
#> [4] severely dependent   severely dependent   severely dependent  
#> 4 Levels: independent slightly dependent ... severely dependent

# structure of numeric values won't be changed
# by this function, it only applies to labelled vectors
# (typically categorical or factor variables)

str(efc$e17age)

#>  num [1:908] 83 88 82 67 84 85 74 87 79 83 ...
#>  - attr(*, "label")= chr "elder' age"

str(as_label(efc$e17age))

#>  num [1:908] 83 88 82 67 84 85 74 87 79 83 ...
#>  - attr(*, "label")= chr "elder' age"

# factor with non-numeric levels
as_label(factor(c("a", "b", "c")))

#> [1] a b c
#> Levels: a b c

# factor with non-numeric levels, prefixed
x <- factor(c("a", "b", "c"))
x <- set_labels(x, labels = c("ape", "bear", "cat"))
as_label(x, prefix = TRUE)

#> [1] [a] ape  [b] bear [c] cat 
#> Levels: [a] ape [b] bear [c] cat

# create vector
x <- c(1, 2, 3, 2, 4, NA)
# add less labels than values
x <- set_labels(
  x,
  labels = c("yes", "maybe", "no"),
  force.labels = FALSE,
  force.values = FALSE
)

#> "x" has more values than "labels", hence not all values are labelled.

# convert to label w/o non-labelled values
as_label(x)

#> [1] yes   maybe no    maybe <NA>  <NA> 
#> Levels: yes maybe no

# convert to label, including non-labelled values
as_label(x, add.non.labelled = TRUE)

#> [1] yes   maybe no    maybe 4     <NA> 
#> Levels: yes maybe no 4

# create labelled integer, with missing flag
library(haven)
x <- labelled(
  c(1:3, tagged_na("a", "c", "z"), 4:1, 2:3),
  c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
    "Refused" = tagged_na("a"), "Not home" = tagged_na("z"))
)

# to labelled factor, with missing labels
as_label(x, drop.na = FALSE)

#>  [1] Agreement    <NA>         <NA>         Refused      First       
#>  [6] Not home     Disagreement <NA>         <NA>         Agreement   
#> [11] <NA>         <NA>        
#> Levels: Agreement Disagreement First Refused Not home

# to labelled factor, missings removed
as_label(x, drop.na = TRUE)

#>  [1] Agreement    <NA>         <NA>         <NA>         <NA>        
#>  [6] <NA>         Disagreement <NA>         <NA>         Agreement   
#> [11] <NA>         <NA>        
#> Levels: Agreement Disagreement

# keep missings, and use non-labelled values as well
as_label(x, add.non.labelled = TRUE, drop.na = FALSE)

#>  [1] Agreement    2            3            Refused      First       
#>  [6] Not home     Disagreement 3            2            Agreement   
#> [11] 2            3           
#> Levels: Agreement 2 3 Disagreement First Refused Not home

# convert labelled character to factor
dummy <- c("M", "F", "F", "X")
dummy <- set_labels(
  dummy,
  labels = c(`M` = "Male", `F` = "Female", `X` = "Refused")
)
get_labels(dummy,, "p")

#> [1] "[F] Female"  "[M] Male"    "[X] Refused"

as_label(dummy)

#> [1] Male    Female  Female  Refused
#> Levels: Female Male Refused

# drop unused factor levels, but preserve variable label
x <- factor(c("a", "b", "c"), levels = c("a", "b", "c", "d"))
x <- set_labels(x, labels = c("ape", "bear", "cat"))
set_label(x) <- "A factor!"
x

#> [1] a b c
#> attr(,"labels")
#>  ape bear  cat 
#>    a    b    c 
#> attr(,"label")
#> [1] A factor!
#> Levels: a b c d

as_label(x, drop.levels = TRUE)

#> [1] ape  bear cat 
#> attr(,"label")
#> [1] A factor!
#> Levels: ape bear cat

# change variable label
as_label(x, var.label = "New variable label!", drop.levels = TRUE)

#> [1] ape  bear cat 
#> attr(,"label")
#> [1] New variable label!
#> Levels: ape bear cat

# convert to numeric and back again, preserving label attributes
# *and* values in numeric vector
x <- c(0, 1, 0, 4)
x <- set_labels(x, labels = c(`null` = 0, `one` = 1, `four` = 4))

# to factor
as_label(x)

#> [1] null one  null four
#> Levels: null one four

# to factor, back to numeric - values are 1, 2 and 3,
# instead of original 0, 1 and 4
as_numeric(as_label(x))

#> [1] 1 2 1 3
#> attr(,"labels")
#> null  one four 
#>    1    2    3 

# preserve label-attributes when converting to factor, use these attributes
# to restore original numeric values when converting back to numeric
as_numeric(as_label(x, keep.labels = TRUE), use.labels = TRUE)

#> [1] 0 1 0 4
#> attr(,"labels")
#> null  one four 
#>    0    1    4 

# easily coerce specific variables in a data frame to factor
# and keep other variables, with their class preserved
as_label(efc, e42dep, e16sex, c172code)

#>    c12hour e15relat e16sex e17age               e42dep c82cop1 c83cop2 c84cop3
#> 1       16        2 female     83 moderately dependent       3       2       2
#> 2      148        2 female     88 moderately dependent       3       3       3
#> 3       70        1 female     82 moderately dependent       2       2       1
#> 4      168        1 female     67   severely dependent       4       1       3
#> 5      168        2 female     84   severely dependent       3       2       1
#> 6       16        2 female     85   severely dependent       2       2       3
#> 7      161        1   male     74   severely dependent       4       2       4
#> 8      110        4 female     87   severely dependent       3       2       2
#> 9       28        2 female     79   severely dependent       3       2       3
#> 10      40        2 female     83   severely dependent       3       2       1
#> 11     100        1   male     68   severely dependent       3       4       4
#> 12      25        8 female     97 moderately dependent       3       3       3
#> 13      25        2 female     80   severely dependent       3       2       2
#> 14      24        1 female     75 moderately dependent       3       2       4
#> 15      56        2 female     82 moderately dependent       2       3       3
#> 16      20        2 female     89 moderately dependent       4       2       1
#> 17      25        1   male     80          independent       3       2       1
#> 18     126        1   male     72 moderately dependent       4       2       1
#> 19     168        2   male     94 moderately dependent       3       2       1
#> 20     118        1   male     79   severely dependent       3       2       4
#> 21     150        2 female     89   severely dependent       3       2       3
#> 22      50        1   male     67 moderately dependent       4       1       2
#> 23      18        2 female     94   severely dependent       3       1       1
#> 24     168        2 female     83   severely dependent       3       2       2
#> 25      15        2 female     85 moderately dependent       3       2       1
#> 26     168        1   male     80   severely dependent       2       3       4
#> 27      12        2 female     88   severely dependent       2       3       3
#> 28       7        1   male     76   slightly dependent       2       2       2
#> 29      35        2 female     84 moderately dependent       3       2       2
#> 30     168        2 female     95   severely dependent       2       2       1
#> 31     150        2   male     88   severely dependent       3       2       4
#> 32     168        1   male     87   severely dependent       3       1       3
#> 33     168        4 female     89 moderately dependent       3       2       2
#> 34     119        1   male     80   severely dependent       2       2       4
#> 35     168        1 female     75   severely dependent       3       2       1
#> 36     168        1   male     82   severely dependent       3       2       1
#> 37     168        1   male     69   severely dependent       3       4       3
#> 38      28        2 female     91   slightly dependent       3       1       1
#>    c85cop4 c86cop5 c87cop6 c88cop7 c89cop8 c90cop9 c160age c161sex
#> 1        2       1       1       2       3       3      56       2
#> 2        3       4       1       3       2       2      54       2
#> 3        4       1       1       1       4       3      80       1
#> 4        1       1       1       1       2       4      69       1
#> 5        2       2       2       1       4       4      47       2
#> 6        3       3       2       2       1       1      56       1
#> 7        1       1       2       4       1       4      61       2
#> 8        1       1       1       2       3       3      67       2
#> 9        2       2       1       3       1       3      59       2
#> 10       2       1       1       1       1       3      49       2
#> 11       4       4       4       4       1       1      66       2
#> 12       1       3       1       4       3       1      47       2
#> 13       2       2       1       2       4       4      58       2
#> 14       4       1       1       2       4       4      75       1
#> 15       3       2       2       1       1       1      49       2
#> 16       3       3       1       2       1       3      56       2
#> 17       2       1       1       2       4       4      75       2
#> 18       2       1       1       2       3       3      70       2
#> 19       2       2       1       3       1       4      52       1
#> 20       2       1       3       3       2       2      48       2
#> 21       2       1       1       1       1       3      58       2
#> 22       1       1       1       1       2       2      65       2
#> 23       1       1       1       1       2       4      49       2
#> 24       4       1       1       2       3       4      60       2
#> 25       2       1       1       2       3       3      55       2
#> 26       4       3       1       4       1       4      62       2
#> 27      NA       1       1       4       2       2      68       2
#> 28       2       1       1       1       3       2      76       2
#> 29       2       4       1       2       2       1      58       2
#> 30       2       1       1       1       2       4      65       2
#> 31       4       4       4       4       1       3      63       1
#> 32       1       1       1       1       1       4      79       2
#> 33       3       1       1       3       1       1      65       2
#> 34       4       2       1       4       2       3      74       2
#> 35       2       1       1       1       3       4      76       1
#> 36       2       1       1       1       4       4      73       2
#> 37       2       1       1       4       1       1      67       2
#> 38       2       1       1       1       4       4      62       2
#>                           c172code c175empl barthtot neg_c_7 pos_v_4 quol_5
#> 1  intermediate level of education        1       75      12      12     14
#> 2  intermediate level of education        1       75      20      11     10
#> 3           low level of education        0       35      11      13      7
#> 4  intermediate level of education        0        0      10      15     12
#> 5  intermediate level of education        0       25      12      15     19
#> 6  intermediate level of education        1       60      19       9      8
#> 7  intermediate level of education        0        5      15      13     20
#> 8  intermediate level of education        0       35      11      14     20
#> 9                             <NA>        0       15      15      13      8
#> 10 intermediate level of education        0        0      10      13     15
#> 11 intermediate level of education        0       25      28       9      1
#> 12 intermediate level of education        1       85      18       8     19
#> 13         high level of education        0       15      13      14     12
#> 14          low level of education        0       70      18      14      8
#> 15         high level of education        1       NA      16       9      8
#> 16 intermediate level of education        0        0      13      14      6
#> 17 intermediate level of education        0       95      11      15     16
#> 18 intermediate level of education        0       55      11      13     14
#> 19         high level of education        1       55      13      13     15
#> 20         high level of education        1       45      17      12     10
#> 21 intermediate level of education        1        0      11      11     15
#> 22 intermediate level of education        0       70       9      12     21
#> 23 intermediate level of education        1       20       8      15     18
#> 24 intermediate level of education        0        0      14      15     15
#> 25 intermediate level of education        0       95      11      14     16
#> 26 intermediate level of education        0       10      23      12      0
#> 27         high level of education        0       75      NA      10     10
#> 28 intermediate level of education        0       90      11       9     14
#> 29 intermediate level of education        0       60      15      11      4
#> 30          low level of education        0       20      11      NA     15
#> 31 intermediate level of education        1       NA      25      13      9
#> 32          low level of education        0       50       9      12     18
#> 33 intermediate level of education        0       75      15       9     12
#> 34 intermediate level of education        0       50      20      12      1
#> 35 intermediate level of education        0       30       9      15      4
#> 36          low level of education        0        0      10      14     13
#> 37 intermediate level of education        0       30      19       9      4
#> 38 intermediate level of education        0       90       8      14     16
#>    resttotn tot_sc_e n4pstu nur_pst
#> 1         0        4      0      NA
#> 2         4        0      0      NA
#> 3         0        1      2       2
#> 4         2        0      3       3
#> 5         2        1      2       2
#> 6         1        3      2       2
#> 7         0        0      3       3
#> 8         0        1      1       1
#> 9         0        2      3       3
#> 10        1        1      3       3
#> 11        1        1      3       3
#> 12        1        1      1       1
#> 13        0        3      3       3
#> 14        0        0      1       1
#> 15        3        3      0      NA
#> 16        0        2      0      NA
#> 17        0        2      0      NA
#> 18        0        0      2       2
#> 19        3        1      1       1
#> 20        0        7      2       2
#> 21        1        1      4       3
#> 22        0        0      1       1
#> 23        1        0      2       2
#> 24        1        1      3       3
#> 25        0        2      1       1
#> 26        1        1      2       2
#> 27        0        5      2       2
#> 28        0        2      0      NA
#> 29        0        2      2       2
#> 30        0        0      3       3
#> 31        1        2      3       3
#> 32        0        1      2       2
#> 33        0        1      1       1
#> 34        0        7      3       3
#> 35        0        4      3       3
#> 36        0        1      3       3
#> 37        0        6      3       3
#> 38        1        1      0      NA
#>  [ reached 'max' / getOption("max.print") -- omitted 870 rows ]