Generic case conversion for labels

This function wraps to_any_case() from the snakecase package with certain defaults for the sep_in and sep_out arguments, used for instance to convert cases in term_labels.

convert_case(lab, case = NULL, verbose = FALSE, ...)

Arguments

lab
Character vector that should be case converted.

case
Desired target case. Labels will automatically converted into the specified character case. See to_any_case() for more details on this argument.

verbose
Toggle warnings and messages on or off.

...
Further arguments passed down to to_any_case(), like sep_in or sep_out.

Value

lab, with converted case.

Details

When calling to_any_case() from snakecase, the sep_in argument is set to "(?<!\\d)\\.", and the sep_out to " ". This gives feasible results from variable labels for plot annotations.

Examples

data(iris)
convert_case(colnames(iris))

#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     

convert_case(colnames(iris), case = "snake")

#> [1] "sepal length" "sepal width"  "petal length" "petal width"  "species"