sjlabelled-package {sjlabelled}	R Documentation

Labelled Data Utility Functions

Description

Purpose of this package

Collection of miscellaneous utility functions (especially intended for people coming from other statistical software packages like 'SPSS', and/or who are new to R), supporting following common tasks when working with labelled data:

Reading and writing data between R and other statistical software packages like 'SPSS', 'SAS' or 'Stata'

Easy ways to get, set and change value and variable label attributes, to convert labelled vectors into factors (and vice versa), or to deal with multiple declared missing values etc.

Author(s)

Daniel Lüdecke d.luedecke@uke.de

[Package sjlabelled version 1.1.3 Index]