Retrieve labels of model terms from regression models

This function retrieves variable labels from model terms. In case of categorical variables, where one variable has multiple dummies, variable name and category value is returned.

term_labels(
  models,
  mark.cat = FALSE,
  case = NULL,
  prefix = c("none", "varname", "label"),
  ...
)

get_term_labels(
  models,
  mark.cat = FALSE,
  case = NULL,
  prefix = c("none", "varname", "label"),
  ...
)

response_labels(models, case = NULL, multi.resp = FALSE, mv = FALSE, ...)

get_dv_labels(models, case = NULL, multi.resp = FALSE, mv = FALSE, ...)

Arguments

models
One or more fitted regression models. May also be glm's or mixed models.

mark.cat
Logical, if TRUE, the returned vector has an attribute with logical values, which indicate whether a label indicates the value from a factor category (attribute value is TRUE) or a term's variable labels (attribute value is FALSE).

case
Desired target case. Labels will automatically converted into the specified character case. See to_any_case() for more details on this argument.

prefix
Indicates whether the value labels of categorical variables should be prefixed, e.g. with the variable name or variable label. May be abbreviated. See 'Examples',

...
Further arguments passed down to to_any_case(), like preprocess or postprocess.

mv, multi.resp
Logical, if TRUE and models is a multivariate response model from a brmsfit object, then the labels for each dependent variable (multiple responses) are returned.

Value

For term_labels(), a (named) character vector with variable labels of all model terms, which can be used, for instance, as axis labels to annotate plots.

For response_labels(), a character vector with variable labels from all dependent variables of models.

Details

Typically, the variable labels from model terms are returned. However, for categorical terms that have estimates for each category, the value labels are returned as well. As the return value is a named vector, you can easily use it with ggplot2's scale_*() functions to annotate plots.

Examples

# use data set with labelled data
data(efc)

fit <- lm(barthtot ~ c160age + c12hour + c161sex + c172code, data = efc)
term_labels(fit)
#>                                    c160age 
#>                               "carer' age" 
#>                                    c12hour 
#> "average number of hours of care per week" 
#>                                    c161sex 
#>                           "carer's gender" 
#>                                   c172code 
#>               "carer's level of education" 

# make "education" categorical
library(sjmisc)
efc$c172code <- to_factor(efc$c172code)
fit <- lm(barthtot ~ c160age + c12hour + c161sex + c172code, data = efc)
term_labels(fit)
#>                                    c160age 
#>                               "carer' age" 
#>                                    c12hour 
#> "average number of hours of care per week" 
#>                                    c161sex 
#>                           "carer's gender" 
#>                                   c172code 
#>               "carer's level of education" 
#>                                  c172code1 
#>                   "low level of education" 
#>                                  c172code2 
#>          "intermediate level of education" 
#>                                  c172code3 
#>                  "high level of education" 

# prefix value of categorical variables with variable name
term_labels(fit, prefix = "varname")
#>                                     c160age 
#>                                "carer' age" 
#>                                     c12hour 
#>  "average number of hours of care per week" 
#>                                     c161sex 
#>                            "carer's gender" 
#>                                    c172code 
#>                "carer's level of education" 
#>                                   c172code1 
#>          "c172code: low level of education" 
#>                                   c172code2 
#> "c172code: intermediate level of education" 
#>                                   c172code3 
#>         "c172code: high level of education" 

# prefix value of categorical variables with value label
term_labels(fit, prefix = "label")
#>                                                       c160age 
#>                                                  "carer' age" 
#>                                                       c12hour 
#>                    "average number of hours of care per week" 
#>                                                       c161sex 
#>                                              "carer's gender" 
#>                                                      c172code 
#>                                  "carer's level of education" 
#>                                                     c172code1 
#>          "carer's level of education: low level of education" 
#>                                                     c172code2 
#> "carer's level of education: intermediate level of education" 
#>                                                     c172code3 
#>         "carer's level of education: high level of education" 

# get label of dv
response_labels(fit)
#> [1] "Total score BARTHEL INDEX"