write_spss {sjlabelled}	R Documentation

Write data to other statistical software packages

Description

These functions write the content of a data frame to an SPSS, SAS or Stata-file.

Usage

write_spss(x, path, drop.na = FALSE)

write_stata(x, path, drop.na = FALSE, version = 14)

write_sas(x, path, drop.na = FALSE)

Arguments

x
A data frame that should be saved as file.

path
File path of the output file.

drop.na
Logical, if TRUE, tagged NA values with value labels will be converted to regular NA's. Else, tagged NA values will be replaced with their value labels. See 'Examples' and get_na.

version
File version to use. Supports versions 8-14.

[Package sjlabelled version 1.1.3 Index]