Drop, add or convert (non-)labelled values

For (partially) labelled vectors, zap_labels() will replace all values that have a value label attribute with NA; zap_unlabelled(), as counterpart, will replace all values that don't have a value label attribute with NA.

drop_labels() drops all value labels for unused values, i.e. values that are not present in a vector. fill_labels() is the counterpart to drop_labels() and adds value labels to a partially labelled vector, i.e. if not all values are labelled, non-labelled values get labels.

drop_labels(x, ..., drop.na = TRUE)

fill_labels(x, ...)

zap_labels(x, ...)

zap_unlabelled(x, ...)

Arguments

x
(partially) labelled vector or a data frame with such vectors.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples'.

drop.na
Logical, whether existing value labels of tagged NA values (see tagged_na) should be removed (drop.na = TRUE, the default) or preserved (drop.na = FALSE). See get_na for more details on tagged NA values.

Value

For zap_labels(), x, where all labelled values are converted to NA.

For zap_unlabelled(), x, where all non-labelled values are converted to NA.

For drop_labels(), x, where value labels for non-existing values are removed.

For fill_labels(), x, where labels for non-labelled values are added.

If x is a data frame, the complete data frame x will be returned, with variables specified in ... being converted; if ... is not specified, applies to all variables in the data frame.

Examples

# zap_labels() ----

data(efc)
str(efc$e42dep)

#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:4] 1 2 3 4
#>   ..- attr(*, "names")= chr [1:4] "independent" "slightly dependent" "moderately dependent" "severely dependent"

x <- set_labels(
  efc$e42dep,
  labels = c("independent" = 1, "severe dependency" = 4)
)
table(x)

#> x
#>   1   2   3   4 
#>  66 225 306 304 

get_values(x)

#> [1] 1 4

str(x)

#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# zap all labelled values
table(zap_labels(x))

#> 
#>   2   3 
#> 225 306 

get_values(zap_labels(x))

#> numeric(0)

str(zap_labels(x))

#>  num [1:908] 3 3 3 NA NA NA NA NA NA NA ...
#>  - attr(*, "labels")= Named num(0) 
#>   ..- attr(*, "names")= chr(0) 

# zap all unlabelled values
table(zap_unlabelled(x))

#> 
#>   1   4 
#>  66 304 

get_values(zap_unlabelled(x))

#> [1] 1 4

str(zap_unlabelled(x))

#>  num [1:908] NA NA NA 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# in a pipe-workflow
library(dplyr)
efc %>%
  select(c172code, e42dep) %>%
  set_labels(
    e42dep,
    labels = c("independent" = 1, "severe dependency" = 4)
  ) %>%
  zap_labels()

#>     c172code e42dep
#> 1         NA      3
#> 2         NA      3
#> 3         NA      3
#> 4         NA     NA
#> 5         NA     NA
#> 6         NA     NA
#> 7         NA     NA
#> 8         NA     NA
#> 9         NA     NA
#> 10        NA     NA
#> 11        NA     NA
#> 12        NA      3
#> 13        NA     NA
#> 14        NA      3
#> 15        NA      3
#> 16        NA      3
#> 17        NA     NA
#> 18        NA      3
#> 19        NA      3
#> 20        NA     NA
#> 21        NA     NA
#> 22        NA      3
#> 23        NA     NA
#> 24        NA     NA
#> 25        NA      3
#> 26        NA     NA
#> 27        NA     NA
#> 28        NA      2
#> 29        NA      3
#> 30        NA     NA
#> 31        NA     NA
#> 32        NA     NA
#> 33        NA      3
#> 34        NA     NA
#> 35        NA     NA
#> 36        NA     NA
#> 37        NA     NA
#> 38        NA      2
#> 39        NA     NA
#> 40        NA      3
#> 41        NA      3
#> 42        NA     NA
#> 43        NA     NA
#> 44        NA     NA
#> 45        NA     NA
#> 46        NA     NA
#> 47        NA     NA
#> 48        NA     NA
#> 49        NA      3
#> 50        NA     NA
#> 51        NA     NA
#> 52        NA     NA
#> 53        NA     NA
#> 54        NA     NA
#> 55        NA     NA
#> 56        NA     NA
#> 57        NA     NA
#> 58        NA     NA
#> 59        NA      3
#> 60        NA     NA
#> 61        NA     NA
#> 62        NA     NA
#> 63        NA     NA
#> 64        NA      3
#> 65        NA      3
#> 66        NA      3
#> 67        NA     NA
#> 68        NA      3
#> 69        NA     NA
#> 70        NA     NA
#> 71        NA     NA
#> 72        NA     NA
#> 73        NA     NA
#> 74        NA     NA
#> 75        NA     NA
#> 76        NA     NA
#> 77        NA     NA
#> 78        NA      3
#> 79        NA      3
#> 80        NA     NA
#> 81        NA      2
#> 82        NA      3
#> 83        NA     NA
#> 84        NA      3
#> 85        NA     NA
#> 86        NA     NA
#> 87        NA     NA
#> 88        NA     NA
#> 89        NA     NA
#> 90        NA     NA
#> 91        NA      3
#> 92        NA     NA
#> 93        NA      2
#> 94        NA      3
#> 95        NA     NA
#> 96        NA      3
#> 97        NA     NA
#> 98        NA     NA
#> 99        NA      3
#> 100       NA     NA
#> 101       NA      3
#> 102       NA      3
#> 103       NA      3
#> 104       NA     NA
#> 105       NA     NA
#> 106       NA      3
#> 107       NA      3
#> 108       NA      2
#> 109       NA      3
#> 110       NA     NA
#> 111       NA     NA
#> 112       NA      3
#> 113       NA     NA
#> 114       NA     NA
#> 115       NA     NA
#> 116       NA      2
#> 117       NA      2
#> 118       NA      3
#> 119       NA     NA
#> 120       NA     NA
#> 121       NA      3
#> 122       NA     NA
#> 123       NA     NA
#> 124       NA     NA
#> 125       NA      3
#> 126       NA     NA
#> 127       NA     NA
#> 128       NA     NA
#> 129       NA      2
#> 130       NA     NA
#> 131       NA      3
#> 132       NA      3
#> 133       NA     NA
#> 134       NA     NA
#> 135       NA      3
#> 136       NA      3
#> 137       NA     NA
#> 138       NA      2
#> 139       NA     NA
#> 140       NA      3
#> 141       NA      2
#> 142       NA     NA
#> 143       NA     NA
#> 144       NA      3
#> 145       NA     NA
#> 146       NA     NA
#> 147       NA     NA
#> 148       NA     NA
#> 149       NA     NA
#> 150       NA      3
#> 151       NA     NA
#> 152       NA     NA
#> 153       NA     NA
#> 154       NA     NA
#> 155       NA     NA
#> 156       NA      3
#> 157       NA      3
#> 158       NA     NA
#> 159       NA      3
#> 160       NA      3
#> 161       NA      3
#> 162       NA      3
#> 163       NA     NA
#> 164       NA     NA
#> 165       NA      3
#> 166       NA     NA
#> 167       NA     NA
#> 168       NA     NA
#> 169       NA      3
#> 170       NA     NA
#> 171       NA      3
#> 172       NA      3
#> 173       NA      3
#> 174       NA      3
#> 175       NA     NA
#> 176       NA     NA
#> 177       NA      3
#> 178       NA     NA
#> 179       NA     NA
#> 180       NA     NA
#> 181       NA     NA
#> 182       NA      3
#> 183       NA      3
#> 184       NA     NA
#> 185       NA     NA
#> 186       NA     NA
#> 187       NA     NA
#> 188       NA     NA
#> 189       NA      2
#> 190       NA      3
#> 191       NA      3
#> 192       NA     NA
#> 193       NA      2
#> 194       NA     NA
#> 195       NA      3
#> 196       NA      3
#> 197       NA      2
#> 198       NA     NA
#> 199       NA      2
#> 200       NA      3
#> 201       NA      3
#> 202       NA     NA
#> 203       NA      3
#> 204       NA     NA
#> 205       NA      3
#> 206       NA      3
#> 207       NA     NA
#> 208       NA     NA
#> 209       NA      3
#> 210       NA      3
#> 211       NA      3
#> 212       NA     NA
#> 213       NA     NA
#> 214       NA     NA
#> 215       NA      2
#> 216       NA      3
#> 217       NA     NA
#> 218       NA     NA
#> 219       NA     NA
#> 220       NA      2
#> 221       NA      3
#> 222       NA     NA
#> 223       NA     NA
#> 224       NA      3
#> 225       NA     NA
#> 226       NA      3
#> 227       NA     NA
#> 228       NA      3
#> 229       NA      3
#> 230       NA      3
#> 231       NA      3
#> 232       NA      3
#> 233       NA     NA
#> 234       NA      3
#> 235       NA     NA
#> 236       NA     NA
#> 237       NA      3
#> 238       NA      2
#> 239       NA      3
#> 240       NA      2
#> 241       NA     NA
#> 242       NA      3
#> 243       NA      3
#> 244       NA      2
#> 245       NA      2
#> 246       NA      3
#> 247       NA     NA
#> 248       NA      2
#> 249       NA      3
#> 250       NA      2
#> 251       NA      3
#> 252       NA      3
#> 253       NA      3
#> 254       NA     NA
#> 255       NA     NA
#> 256       NA      3
#> 257       NA      3
#> 258       NA     NA
#> 259       NA      2
#> 260       NA      2
#> 261       NA      2
#> 262       NA     NA
#> 263       NA     NA
#> 264       NA      3
#> 265       NA      3
#> 266       NA     NA
#> 267       NA      3
#> 268       NA      3
#> 269       NA      3
#> 270       NA      3
#> 271       NA      2
#> 272       NA      3
#> 273       NA      2
#> 274       NA      2
#> 275       NA      3
#> 276       NA      3
#> 277       NA      3
#> 278       NA      3
#> 279       NA      2
#> 280       NA      3
#> 281       NA      3
#> 282       NA      2
#> 283       NA      2
#> 284       NA      2
#> 285       NA      3
#> 286       NA      3
#> 287       NA     NA
#> 288       NA     NA
#> 289       NA     NA
#> 290       NA     NA
#> 291       NA      2
#> 292       NA      2
#> 293       NA      3
#> 294       NA     NA
#> 295       NA      3
#> 296       NA      2
#> 297       NA      3
#> 298       NA      3
#> 299       NA      2
#> 300       NA      2
#> 301       NA     NA
#> 302       NA     NA
#> 303       NA      3
#> 304       NA      2
#> 305       NA      2
#> 306       NA      3
#> 307       NA     NA
#> 308       NA      2
#> 309       NA      3
#> 310       NA     NA
#> 311       NA      3
#> 312       NA      3
#> 313       NA      3
#> 314       NA     NA
#> 315       NA      3
#> 316       NA      3
#> 317       NA     NA
#> 318       NA     NA
#> 319       NA      3
#> 320       NA      2
#> 321       NA      3
#> 322       NA      3
#> 323       NA      2
#> 324       NA     NA
#> 325       NA      2
#> 326       NA      3
#> 327       NA      3
#> 328       NA      3
#> 329       NA     NA
#> 330       NA     NA
#> 331       NA      2
#> 332       NA     NA
#> 333       NA     NA
#> 334       NA     NA
#> 335       NA      3
#> 336       NA      2
#> 337       NA      2
#> 338       NA      3
#> 339       NA      3
#> 340       NA      2
#> 341       NA      2
#> 342       NA      2
#> 343       NA      2
#> 344       NA      2
#> 345       NA      3
#> 346       NA     NA
#> 347       NA      3
#> 348       NA     NA
#> 349       NA     NA
#> 350       NA      2
#> 351       NA     NA
#> 352       NA      2
#> 353       NA      2
#> 354       NA      2
#> 355       NA      3
#> 356       NA      3
#> 357       NA     NA
#> 358       NA      3
#> 359       NA      3
#> 360       NA      2
#> 361       NA      3
#> 362       NA      3
#> 363       NA      2
#> 364       NA      3
#> 365       NA      2
#> 366       NA     NA
#> 367       NA     NA
#> 368       NA      2
#> 369       NA      2
#> 370       NA      2
#> 371       NA      2
#> 372       NA     NA
#> 373       NA     NA
#> 374       NA     NA
#> 375       NA      2
#> 376       NA      2
#> 377       NA      2
#> 378       NA      3
#> 379       NA     NA
#> 380       NA     NA
#> 381       NA     NA
#> 382       NA      3
#> 383       NA      3
#> 384       NA     NA
#> 385       NA      3
#> 386       NA      2
#> 387       NA      2
#> 388       NA     NA
#> 389       NA      3
#> 390       NA     NA
#> 391       NA     NA
#> 392       NA      3
#> 393       NA     NA
#> 394       NA      3
#> 395       NA      2
#> 396       NA     NA
#> 397       NA      2
#> 398       NA      2
#> 399       NA     NA
#> 400       NA      2
#> 401       NA     NA
#> 402       NA     NA
#> 403       NA     NA
#> 404       NA      2
#> 405       NA     NA
#> 406       NA      2
#> 407       NA     NA
#> 408       NA      2
#> 409       NA     NA
#> 410       NA      3
#> 411       NA     NA
#> 412       NA     NA
#> 413       NA     NA
#> 414       NA      2
#> 415       NA      2
#> 416       NA      3
#> 417       NA      2
#> 418       NA     NA
#> 419       NA     NA
#> 420       NA      3
#> 421       NA     NA
#> 422       NA      3
#> 423       NA     NA
#> 424       NA      3
#> 425       NA     NA
#> 426       NA     NA
#> 427       NA      2
#> 428       NA     NA
#> 429       NA     NA
#> 430       NA     NA
#> 431       NA     NA
#> 432       NA      3
#> 433       NA      2
#> 434       NA      2
#> 435       NA      3
#> 436       NA      2
#> 437       NA      3
#> 438       NA     NA
#> 439       NA     NA
#> 440       NA      3
#> 441       NA      3
#> 442       NA      2
#> 443       NA      3
#> 444       NA      3
#> 445       NA      2
#> 446       NA     NA
#> 447       NA      3
#> 448       NA     NA
#> 449       NA      3
#> 450       NA      3
#> 451       NA      3
#> 452       NA      2
#> 453       NA      3
#> 454       NA      3
#> 455       NA      3
#> 456       NA     NA
#> 457       NA      2
#> 458       NA     NA
#> 459       NA      3
#> 460       NA      3
#> 461       NA      3
#> 462       NA      2
#> 463       NA     NA
#> 464       NA      3
#> 465       NA     NA
#> 466       NA     NA
#> 467       NA     NA
#> 468       NA     NA
#> 469       NA      3
#> 470       NA      2
#> 471       NA     NA
#> 472       NA     NA
#> 473       NA     NA
#> 474       NA     NA
#> 475       NA     NA
#> 476       NA     NA
#> 477       NA     NA
#> 478       NA      2
#> 479       NA     NA
#> 480       NA      3
#> 481       NA     NA
#> 482       NA      3
#> 483       NA     NA
#> 484       NA      2
#> 485       NA      3
#> 486       NA      3
#> 487       NA      3
#> 488       NA     NA
#> 489       NA     NA
#> 490       NA      3
#> 491       NA      2
#> 492       NA      3
#> 493       NA     NA
#> 494       NA      2
#> 495       NA      3
#> 496       NA      2
#> 497       NA      2
#> 498       NA      3
#> 499       NA      3
#> 500       NA      2
#>  [ reached 'max' / getOption("max.print") -- omitted 408 rows ]

# drop_labels() ----

library(sjmisc)
rp <- rec_pattern(1, 100)
rp

#> $pattern
#> [1] "1:5=1;6:10=2;11:15=3;16:20=4;21:25=5;26:30=6;31:35=7;36:40=8;41:45=9;46:50=10;51:55=11;56:60=12;61:65=13;66:70=14;71:75=15;76:80=16;81:85=17;86:90=18;91:95=19;96:100=20;"
#> 
#> $labels
#>        1        2        3        4        5        6        7        8 
#>    "1-5"   "6-10"  "11-15"  "16-20"  "21-25"  "26-30"  "31-35"  "36-40" 
#>        9       10       11       12       13       14       15       16 
#>  "41-45"  "46-50"  "51-55"  "56-60"  "61-65"  "66-70"  "71-75"  "76-80" 
#>       17       18       19       20 
#>  "81-85"  "86-90"  "91-95" "96-100" 
#> 

# sample data
data(efc)
# recode carers age into groups of width 5
x <- rec(efc$c160age, rec = rp$pattern)
# add value labels to new vector
x <- set_labels(x, labels = rp$labels)

# watch result. due to recode-pattern, we have age groups with
# no observations (zero-counts)
frq(x)

#> 
#> carer' age (x) <numeric>
#> # total N=908  valid N=901  mean=11.07  sd=2.69
#> 
#>  val  label frq raw.prc valid.prc cum.prc
#>    1    1-5   0    0.00      0.00    0.00
#>    2   6-10   0    0.00      0.00    0.00
#>    3  11-15   0    0.00      0.00    0.00
#>    4  16-20   5    0.55      0.55    0.55
#>    5  21-25  15    1.65      1.66    2.22
#>    6  26-30  28    3.08      3.11    5.33
#>    7  31-35  50    5.51      5.55   10.88
#>    8  36-40  51    5.62      5.66   16.54
#>    9  41-45  99   10.90     10.99   27.52
#>   10  46-50 126   13.88     13.98   41.51
#>   11  51-55 116   12.78     12.87   54.38
#>   12  56-60 122   13.44     13.54   67.92
#>   13  61-65 123   13.55     13.65   81.58
#>   14  66-70  82    9.03      9.10   90.68
#>   15  71-75  41    4.52      4.55   95.23
#>   16  76-80  32    3.52      3.55   98.78
#>   17  81-85   9    0.99      1.00   99.78
#>   18  86-90   2    0.22      0.22  100.00
#>   19  91-95   0    0.00      0.00  100.00
#>   20 96-100   0    0.00      0.00  100.00
#>   NA   <NA>   7    0.77        NA      NA
#> 

# now, let's drop zero's
frq(drop_labels(x))

#> 
#> carer' age (x) <numeric>
#> # total N=908  valid N=901  mean=11.07  sd=2.69
#> 
#>  val label frq raw.prc valid.prc cum.prc
#>    4 16-20   5    0.55      0.55    0.55
#>    5 21-25  15    1.65      1.66    2.22
#>    6 26-30  28    3.08      3.11    5.33
#>    7 31-35  50    5.51      5.55   10.88
#>    8 36-40  51    5.62      5.66   16.54
#>    9 41-45  99   10.90     10.99   27.52
#>   10 46-50 126   13.88     13.98   41.51
#>   11 51-55 116   12.78     12.87   54.38
#>   12 56-60 122   13.44     13.54   67.92
#>   13 61-65 123   13.55     13.65   81.58
#>   14 66-70  82    9.03      9.10   90.68
#>   15 71-75  41    4.52      4.55   95.23
#>   16 76-80  32    3.52      3.55   98.78
#>   17 81-85   9    0.99      1.00   99.78
#>   18 86-90   2    0.22      0.22  100.00
#>   NA  <NA>   7    0.77        NA      NA
#> 

# drop labels, also drop NA value labels, then also zap tagged NA
library(haven)
x <- labelled(c(1:3, tagged_na("z"), 4:1),
              c("Agreement" = 1, "Disagreement" = 4, "Unused" = 5,
                "Not home" = tagged_na("z")))
x

#> <Labelled double>
#> [1]     1     2     3 NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>      5       Unused
#>  NA(z)     Not home

drop_labels(x, drop.na = FALSE)

#> <Labelled double>
#> [1]     1     2     3 NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>  NA(z)     Not home

drop_labels(x)

#> <Labelled double>
#> [1]     1     2     3 NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement

zap_na_tags(drop_labels(x))

#> <Labelled double>
#> [1]  1  2  3 NA  4  3  2  1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement

# fill_labels() ----

# create labelled integer, with tagged missings
library(haven)
x <- labelled(
  c(1:3, tagged_na("a", "c", "z"), 4:1),
  c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
    "Refused" = tagged_na("a"), "Not home" = tagged_na("z"))
  )
# get current values and labels
x

#> <Labelled double>
#>  [1]     1     2     3 NA(a) NA(c) NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>  NA(c)        First
#>  NA(a)      Refused
#>  NA(z)     Not home

get_labels(x)

#> [1] "Agreement"    "Disagreement"

fill_labels(x)

#> <Labelled double>
#>  [1]     1     2     3 NA(a) NA(c) NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      2            2
#>      3            3
#>      4 Disagreement
#>  NA(c)        First
#>  NA(a)      Refused
#>  NA(z)     Not home

get_labels(fill_labels(x))

#> [1] "Agreement"    "2"            "3"            "Disagreement"

# same as
get_labels(x, non.labelled = TRUE)

#> [1] "Agreement"    "2"            "3"            "Disagreement"