Convert tagged NA values into regular NA

Replaces all tagged_na values with regular NA.

zap_na_tags(x, ...)

Arguments

x
A labelled vector with tagged_na values, or a data frame with such vectors.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples'.

Value

x, where all tagged_na values are converted to NA.

Examples

library(haven)
x <- labelled(
  c(1:3, tagged_na("a", "c", "z"), 4:1),
  c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
    "Refused" = tagged_na("a"), "Not home" = tagged_na("z"))
)
# get current NA values
x

#> <Labelled double>
#>  [1]     1     2     3 NA(a) NA(c) NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>  NA(c)        First
#>  NA(a)      Refused
#>  NA(z)     Not home

get_na(x)

#>    First  Refused Not home 
#>       NA       NA       NA 

zap_na_tags(x)

#> <Labelled double>
#>  [1]  1  2  3 NA NA NA  4  3  2  1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement

get_na(zap_na_tags(x))

#> NULL

# also works with non-labelled vector that have tagged NA values
x <- c(1:5, tagged_na("a"), tagged_na("z"), NA)
haven::print_tagged_na(x)

#> [1]     1     2     3     4     5 NA(a) NA(z)    NA

haven::print_tagged_na(zap_na_tags(x))

#> [1]  1  2  3  4  5 NA NA NA