Convert variable into factor and keep value labels

This function converts a variable into a factor, but preserves variable and value label attributes.

as_factor(x, ...)

to_factor(x, ...)

# S3 method for data.frame
as_factor(x, ..., add.non.labelled = FALSE)

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select-helpers. See 'Examples'.

add.non.labelled
Logical, if TRUE, non-labelled values also get value labels.

Value

A factor, including variable and value labels. If x is a data frame, the complete data frame x will be returned, where variables specified in ... are coerced to factors (including variable and value labels); if ... is not specified, applies to all variables in the data frame.

Details

as_factor converts numeric values into a factor with numeric levels. as_label, however, converts a vector into a factor and uses value labels as factor levels.

Note

This function is intended for use with vectors that have value and variable label attributes. Unlike as.factor, as_factor converts a variable into a factor and preserves the value and variable label attributes.

Adding label attributes is automatically done by importing data sets with one of the read_*-functions, like read_spss. Else, value and variable labels can be manually added to vectors with set_labels and set_label.

Examples

library(sjmisc)
library(magrittr)
data(efc)
# normal factor conversion, loses value attributes
x <- as.factor(efc$e42dep)
frq(x)
#> 
#> x <categorical>
#> # total N=908  valid N=901  mean=2.94  sd=0.94
#> 
#> Value |   N | Raw % | Valid % | Cum. %
#> --------------------------------------
#>     1 |  66 |  7.27 |    7.33 |   7.33
#>     2 | 225 | 24.78 |   24.97 |  32.30
#>     3 | 306 | 33.70 |   33.96 |  66.26
#>     4 | 304 | 33.48 |   33.74 | 100.00
#>  <NA> |   7 |  0.77 |    <NA> |   <NA>
#> 

# factor conversion, which keeps value attributes
x <- as_factor(efc$e42dep)
frq(x)
#> 
#> x <categorical>
#> # total N=908  valid N=901  mean=2.94  sd=0.94
#> 
#> Value |   N | Raw % | Valid % | Cum. %
#> --------------------------------------
#>     1 |  66 |  7.27 |    7.33 |   7.33
#>     2 | 225 | 24.78 |   24.97 |  32.30
#>     3 | 306 | 33.70 |   33.96 |  66.26
#>     4 | 304 | 33.48 |   33.74 | 100.00
#>  <NA> |   7 |  0.77 |    <NA> |   <NA>
#> 

# create parially labelled vector
x <- set_labels(
  efc$e42dep,
  labels = c(
    `1` = "independent",
    `4` = "severe dependency",
    `9` = "missing value"
 ))

# only copy existing value labels
as_factor(x) %>% head()
#> [1] 3 3 3 4 4 4
#> Levels: 1 2 3 4
get_labels(as_factor(x), values = "p")
#> [1] "1" "2" "3" "4"

# also add labels to non-labelled values
as_factor(x, add.non.labelled = TRUE) %>% head()
#> Error: 1 components of `...` were not used.
#> 
#> We detected these problematic arguments:
#> * `add.non.labelled`
#> 
#> Did you misspecify an argument?
get_labels(as_factor(x, add.non.labelled = TRUE), values = "p")
#> Error: 1 components of `...` were not used.
#> 
#> We detected these problematic arguments:
#> * `add.non.labelled`
#> 
#> Did you misspecify an argument?

# easily coerce specific variables in a data frame to factor
# and keep other variables, with their class preserved
as_factor(efc, e42dep, e16sex, c172code) %>% head()
#> Error: 3 components of `...` were not used.
#> 
#> We detected these problematic arguments:
#> * `..1`
#> * `..2`
#> * `..3`
#> 
#> Did you misspecify an argument?

# use select-helpers from dplyr-package
library(dplyr)
as_factor(efc, contains("cop"), c161sex:c175empl) %>% head()
#> Error: 2 components of `...` were not used.
#> 
#> We detected these problematic arguments:
#> * `..1`
#> * `..2`
#> 
#> Did you misspecify an argument?