remove_all_labels {sjlabelled}	R Documentation

Remove value and variable labels from vector or data frame

Description

This function removes value and variable label attributes from a vector or data frame. These attributes are typically added to variables when importing foreign data (see read_spss) or manually adding label attributes with set_labels.

Usage

remove_all_labels(x)

Arguments

x
Vector or data.frame with variable and/or value label attributes

Value

x with removed value and variable label attributes.

See Also

See vignette Labelled Data and the sjlabelled-Package, and copy_labels for adding label attributes (subsetted) data frames.

Examples

data(efc)
str(efc)
str(remove_all_labels(efc))

[Package sjlabelled version 1.1.3 Index]