tidy_labels {sjlabelled}	R Documentation

Repair value labels

Description

Duplicated value labels in variables may cause troubles when saving labelled data, or computing cross tabs (cf. flat_table or plot_xtab). tidy_labels() repairs duplicated value labels by suffixing them with the associated value.

Usage

tidy_labels(x, ..., sep = "_", remove = FALSE)

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples'.

sep
String that will be used to separate the suffixed value from the old label when creating the new value label.

remove
Logical, if TRUE, the original, duplicated value label will be replaced by the value (i.e. the value is not the suffix of the value label, but will become the value label itself). The sep-argument will be ignored in such cases.

Value

x, with "repaired" (unique) value labels for each variable.

Examples

library(sjmisc)
set.seed(123)
x <- set_labels(
  sample(1:5, size = 20, replace = TRUE),
  labels = c("low" = 1, ".." = 2, ".." = 3, ".." = 4, "high" = 5)
)
frq(x)

z <- tidy_labels(x)
frq(z)

z <- tidy_labels(x, sep = ".")
frq(z)

z <- tidy_labels(x, remove = TRUE)
frq(z)

[Package sjlabelled version 1.1.3 Index]