unlabel {sjlabelled}	R Documentation

Convert labelled vectors into normal classes

Description

This function converts labelled class vectors into a generic data format, which means that simply all labelled class attributes will be removed, so all vectors / variables will most likely become atomic.

Usage

unlabel(x, verbose = FALSE)

Arguments

x
A data frame, which contains labelled class vectors or a single vector of class labelled.

verbose
Logical, if TRUE, a progress bar is displayed that indicates the progress of converting the imported data.

Value

A data frame or single vector (depending on x) with common object classes.

Note

This function is currently only used to avoid possible compatibility issues with labelled class vectors. Some known issues with labelled class vectors have already been fixed, so it might be that this function will become redundant in the future.

[Package sjlabelled version 1.1.3 Index]