Drop, add or convert (non-)labelled values

For (partially) labelled vectors, zap_labels() will replace all values that have a value label attribute with NA; zap_unlabelled(), as counterpart, will replace all values that don't have a value label attribute with NA.

drop_labels() drops all value labels for unused values, i.e. values that are not present in a vector. fill_labels() is the counterpart to drop_labels() and adds value labels to a partially labelled vector, i.e. if not all values are labelled, non-labelled values get labels.

drop_labels(x, ..., drop.na = TRUE)

fill_labels(x, ...)

zap_labels(x, ...)

zap_unlabelled(x, ...)

Arguments

x
(partially) labelled() vector or a data frame with such vectors.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select-helpers. See 'Examples'.

drop.na
Logical, whether existing value labels of tagged NA values (see tagged_na) should be removed (drop.na = TRUE, the default) or preserved (drop.na = FALSE). See get_na for more details on tagged NA values.

Value

For zap_labels(), x, where all labelled values are converted to NA.

For zap_unlabelled(), x, where all non-labelled values are converted to NA.

For drop_labels(), x, where value labels for non-existing values are removed.

For fill_labels(), x, where labels for non-labelled values are added.

If x is a data frame, the complete data frame x will be returned, with variables specified in ... being converted; if ... is not specified, applies to all variables in the data frame.

Examples

# zap_labels() ----

data(efc)
str(efc$e42dep)
#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:4] 1 2 3 4
#>   ..- attr(*, "names")= chr [1:4] "independent" "slightly dependent" "moderately dependent" "severely dependent"

x <- set_labels(
  efc$e42dep,
  labels = c("independent" = 1, "severe dependency" = 4)
)
table(x)
#> x
#>   1   2   3   4 
#>  66 225 306 304 
get_values(x)
#> [1] 1 4
str(x)
#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# zap all labelled values
table(zap_labels(x))
#> 
#>   1   2   3   4 
#>  66 225 306 304 
get_values(zap_labels(x))
#> [1] 1 4
str(zap_labels(x))
#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# zap all unlabelled values
table(zap_unlabelled(x))
#> 
#>   1   4 
#>  66 304 
get_values(zap_unlabelled(x))
#> [1] 1 4
str(zap_unlabelled(x))
#>  num [1:908] NA NA NA 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# in a pipe-workflow
library(dplyr)
efc %>%
  select(c172code, e42dep) %>%
  set_labels(
    e42dep,
    labels = c("independent" = 1, "severe dependency" = 4)
  ) %>%
  zap_labels()
#>     c172code e42dep
#> 1          2      3
#> 2          2      3
#> 3          1      3
#> 4          2      4
#> 5          2      4
#> 6          2      4
#> 7          2      4
#> 8          2      4
#> 9         NA      4
#> 10         2      4
#> 11         2      4
#> 12         2      3
#> 13         3      4
#> 14         1      3
#> 15         3      3
#> 16         2      3
#> 17         2      1
#> 18         2      3
#> 19         3      3
#> 20         3      4
#> 21         2      4
#> 22         2      3
#> 23         2      4
#> 24         2      4
#> 25         2      3
#> 26         2      4
#> 27         3      4
#> 28         2      2
#> 29         2      3
#> 30         1      4
#> 31         2      4
#> 32         1      4
#> 33         2      3
#> 34         2      4
#> 35         2      4
#> 36         1      4
#> 37         2      4
#> 38         2      2
#> 39         2      4
#> 40         2      3
#> 41         3      3
#> 42         2      4
#> 43         3      4
#> 44         2      4
#> 45         2      4
#> 46         2      4
#> 47         2      4
#> 48        NA      4
#> 49        NA      3
#> 50         2      4
#> 51         2      4
#> 52         2      4
#> 53         2      4
#> 54         2      4
#> 55         3      4
#> 56         2      4
#> 57         2      4
#> 58         3      4
#> 59         2      3
#> 60         2      4
#> 61        NA      4
#> 62         2      4
#> 63        NA      4
#> 64         3      3
#> 65         2      3
#> 66         3      3
#> 67         3      4
#> 68         2      3
#> 69         2      4
#> 70        NA      4
#> 71         2      4
#> 72         2      4
#> 73         2      4
#> 74         3      4
#> 75         3      4
#> 76         2      4
#> 77         1      4
#> 78         2      3
#> 79         3      3
#> 80         2      4
#> 81         2      2
#> 82         2      3
#> 83         2      4
#> 84         2      3
#> 85         2      4
#> 86         2      4
#> 87         2      4
#> 88         2      4
#> 89         3      4
#> 90         2      4
#> 91         1      3
#> 92         2      1
#> 93         2      2
#> 94        NA      3
#> 95        NA      4
#> 96         2      3
#> 97        NA     NA
#> 98        NA      4
#> 99         1      3
#> 100        2      4
#> 101       NA      3
#> 102        2      3
#> 103        1      3
#> 104       NA      4
#> 105        2      4
#> 106        2      3
#> 107        1      3
#> 108       NA      2
#> 109        2      3
#> 110        2      4
#> 111        2      4
#> 112        2      3
#> 113        3      4
#> 114        2      4
#> 115        2      4
#> 116        1      2
#> 117        2      2
#> 118        2      3
#> 119        1      4
#> 120        1      4
#> 121        3      3
#> 122        1      4
#> 123        1      4
#> 124        2      4
#> 125        2      3
#> 126        2      4
#> 127        2      1
#> 128        2      4
#> 129        3      2
#> 130        1      4
#> 131        1      3
#> 132        2      3
#> 133        1      4
#> 134        1      4
#> 135        2      3
#> 136        3      3
#> 137        1      4
#> 138        2      2
#> 139        1      4
#> 140        2      3
#> 141        2      2
#> 142       NA      4
#> 143        2      4
#> 144        2      3
#> 145        2      4
#> 146        3      4
#> 147       NA      4
#> 148       NA      4
#> 149        1      4
#> 150        2      3
#> 151        2      4
#> 152        2      4
#> 153        3      4
#> 154        3      4
#> 155        1      4
#> 156        2      3
#> 157        3      3
#> 158        3      4
#> 159        3      3
#> 160        3      3
#> 161        3      3
#> 162        3      3
#> 163       NA      4
#> 164        2      4
#> 165        3      3
#> 166        3      4
#> 167       NA      4
#> 168        2      4
#> 169        2      3
#> 170        3      4
#> 171        3      3
#> 172        2      3
#> 173        2      3
#> 174        1      3
#> 175        2      4
#> 176        3      4
#> 177        2      3
#> 178       NA      4
#> 179        3      4
#> 180        2      4
#> 181        1      4
#> 182        3      3
#> 183        2      3
#> 184       NA      4
#> 185        1      4
#> 186        2      4
#> 187        3      4
#> 188        3      4
#> 189        1      2
#> 190        2      3
#> 191        3      3
#> 192        3      4
#> 193       NA      2
#> 194        2      4
#> 195        2      3
#> 196        1      3
#> 197        3      2
#> 198        3      4
#> 199        1      2
#> 200       NA      3
#> 201       NA      3
#> 202        2      4
#> 203        3      3
#> 204        2      4
#> 205        1      3
#> 206        2      3
#> 207        2      4
#> 208        2      4
#> 209        1      3
#> 210       NA      3
#> 211        3      3
#> 212        3      4
#> 213        3      4
#> 214        3      4
#> 215        3      2
#> 216        2      3
#> 217        2      4
#> 218        2      4
#> 219        1      4
#> 220        3      2
#> 221        2      3
#> 222        3      4
#> 223        2      4
#> 224        2      3
#> 225        1      4
#> 226        3      3
#> 227        1      4
#> 228        2      3
#> 229        3      3
#> 230        2      3
#> 231        3      3
#> 232       NA      3
#> 233       NA      4
#> 234       NA      3
#> 235       NA      4
#> 236       NA      4
#> 237        2      3
#> 238        2      2
#> 239        2      3
#> 240        2      2
#> 241        2      1
#> 242        2      3
#> 243        1      3
#> 244       NA      2
#> 245        2      2
#> 246        2      3
#> 247        2      1
#> 248        2      2
#> 249        1      3
#> 250       NA      2
#> 251       NA      3
#> 252        1      3
#> 253       NA      3
#> 254        2      4
#> 255        2      1
#> 256        2      3
#> 257        2      3
#> 258        1      4
#> 259        2      2
#> 260        3      2
#> 261        2      2
#> 262        3      4
#> 263        2      4
#> 264        1      3
#> 265        2      3
#> 266       NA      4
#> 267        2      3
#> 268        3      3
#> 269        2      3
#> 270        1      3
#> 271        1      2
#> 272        1      3
#> 273        2      2
#> 274        1      2
#> 275        2      3
#> 276        1      3
#> 277       NA      3
#> 278       NA      3
#> 279        1      2
#> 280        3      3
#> 281        1      3
#> 282        2      2
#> 283        2      2
#> 284        2      2
#> 285        2      3
#> 286        2      3
#> 287        2      4
#> 288        3      4
#> 289        2      4
#> 290        2      4
#> 291        2      2
#> 292        1      2
#> 293        2      3
#> 294        1      4
#> 295        1      3
#> 296        1      2
#> 297        2      3
#> 298        2      3
#> 299        3      2
#> 300        2      2
#> 301        1      4
#> 302        2      4
#> 303        2      3
#> 304        2      2
#> 305        2      2
#> 306        2      3
#> 307        2      4
#> 308        2      2
#> 309        3      3
#> 310        1      1
#> 311        2      3
#> 312        3      3
#> 313        2      3
#> 314        3      4
#> 315        2      3
#> 316        2      3
#> 317        2      4
#> 318        2      4
#> 319        2      3
#> 320        1      2
#> 321       NA      3
#> 322        2      3
#> 323        3      2
#> 324        3      4
#> 325        3      2
#> 326        1      3
#> 327        2      3
#> 328        1      3
#> 329        2      1
#> 330        2      1
#> 331        2      2
#> 332        2      4
#> 333        2      4
#> 334        1      4
#> 335        2      3
#> 336        3      2
#> 337        1      2
#> 338        2      3
#> 339        3      3
#> 340        2      2
#> 341        2      2
#> 342        2      2
#> 343        2      2
#> 344        2      2
#> 345        2      3
#> 346        2      4
#> 347        2      3
#> 348        3      4
#> 349        2      1
#> 350        3      2
#> 351        3      4
#> 352        1      2
#> 353        2      2
#> 354        1      2
#> 355        1      3
#> 356        1      3
#> 357        2      4
#> 358        2      3
#> 359        1      3
#> 360        1      2
#> 361        1      3
#> 362        2      3
#> 363        1      2
#> 364        1      3
#> 365        2      2
#> 366        1      1
#> 367        2      1
#> 368       NA      2
#> 369       NA      2
#> 370        2      2
#> 371        2      2
#> 372        2      4
#> 373        2      4
#> 374        2      1
#> 375        2      2
#> 376        1      2
#> 377        2      2
#> 378        2      3
#> 379        2      4
#> 380        2      4
#> 381        2      4
#> 382        2      3
#> 383        2      3
#> 384        2      4
#> 385        1      3
#> 386        2      2
#> 387        2      2
#> 388       NA      4
#> 389        1      3
#> 390        2      4
#> 391        2      4
#> 392        3      3
#> 393        1      1
#> 394        3      3
#> 395        2      2
#> 396        2      4
#> 397        3      2
#> 398        3      2
#> 399        1      4
#> 400        3      2
#> 401        1      4
#> 402        1      4
#> 403        2      4
#> 404        2      2
#> 405        2      1
#> 406        2      2
#> 407        1      4
#> 408        1      2
#> 409        2      4
#> 410        2      3
#> 411        2      4
#> 412        2      1
#> 413        1      1
#> 414        2      2
#> 415        2      2
#> 416        3      3
#> 417        3      2
#> 418        3      4
#> 419        3      4
#> 420        2      3
#> 421        1      4
#> 422        2      3
#> 423        3      4
#> 424        2      3
#> 425        2      4
#> 426        2      4
#> 427        3      2
#> 428        2      4
#> 429        2      4
#> 430        2      4
#> 431        3      1
#> 432        3      3
#> 433        2      2
#> 434        1      2
#> 435        2      3
#> 436        2      2
#> 437        2      3
#> 438        3      4
#> 439       NA      4
#> 440        1      3
#> 441        1      3
#> 442        2      2
#> 443        2      3
#> 444        3      3
#> 445        1      2
#> 446        2      4
#> 447        2      3
#> 448        2      4
#> 449        2      3
#> 450        1      3
#> 451        2      3
#> 452        2      2
#> 453        3      3
#> 454        2      3
#> 455        2      3
#> 456        3      4
#> 457        1      2
#> 458        1      4
#> 459        3      3
#> 460        3      3
#> 461        3      3
#> 462        2      2
#> 463        2      4
#> 464        1      3
#> 465        2      4
#> 466        2      4
#> 467        1      4
#> 468        1      1
#> 469        2      3
#> 470        2      2
#> 471        1      4
#> 472        2      1
#> 473        2      1
#> 474        2      1
#> 475        1      1
#> 476        2      1
#> 477        1      4
#> 478        2      2
#> 479        2      1
#> 480        2      3
#> 481        1      4
#> 482        3      3
#> 483        3      4
#> 484        2      2
#> 485        3      3
#> 486        2      3
#> 487       NA      3
#> 488       NA      1
#> 489       NA      1
#> 490        1      3
#> 491        1      2
#> 492        1      3
#> 493        1      4
#> 494        1      2
#> 495        1      3
#> 496        2      2
#> 497        1      2
#> 498        1      3
#> 499        1      3
#> 500        3      2
#> 501        1      4
#> 502        2      4
#> 503        1      3
#> 504        2      4
#> 505        2      4
#> 506        2      4
#> 507        2      3
#> 508        2      3
#> 509        2      3
#> 510        2      3
#> 511        2      2
#> 512        2      3
#> 513        2      2
#> 514        1      2
#> 515        1      3
#> 516        2      3
#> 517        2      4
#> 518       NA      4
#> 519        1      4
#> 520        2      2
#> 521        3      2
#> 522        3      4
#> 523        1      3
#> 524        2      2
#> 525        2      2
#> 526        2      3
#> 527        2      1
#> 528        2      2
#> 529        2      3
#> 530        2      2
#> 531        2      4
#> 532        2      1
#> 533        2      4
#> 534        2      3
#> 535        2      2
#> 536        3      2
#> 537        3      3
#> 538        2      4
#> 539        2      3
#> 540        2      4
#> 541        2      4
#> 542        2      2
#> 543        2      3
#> 544        2      2
#> 545        3      3
#> 546        2      3
#> 547        2      2
#> 548        1      4
#> 549        1      3
#> 550        1      2
#> 551        2      3
#> 552        2      2
#> 553        2      2
#> 554        2      3
#> 555        2      3
#> 556        2      3
#> 557        1      4
#> 558        1      4
#> 559        1      4
#> 560        1      2
#> 561        1      4
#> 562        2      2
#> 563        1      4
#> 564        2      2
#> 565        3      3
#> 566        2      1
#> 567        2      2
#> 568        1      3
#> 569        1      4
#> 570        3      4
#> 571        2      4
#> 572        1      4
#> 573        2      4
#> 574        1      4
#> 575        1      4
#> 576        2      3
#> 577        2      4
#> 578        3      2
#> 579       NA      4
#> 580        2      3
#> 581       NA      4
#> 582        1      3
#> 583        3      3
#> 584        1      3
#> 585        2      2
#> 586        2      2
#> 587        2      3
#> 588        2      4
#> 589        2      4
#> 590        1      4
#> 591        1      1
#> 592        2      4
#> 593        1      4
#> 594        1      3
#> 595        2      4
#> 596        2      4
#> 597        2      4
#> 598        1      2
#> 599        1      4
#> 600        2      4
#> 601        3      3
#> 602        1      4
#> 603        2      3
#> 604        2      2
#> 605        3      2
#> 606        3      2
#> 607        2      2
#> 608        1      1
#> 609        3      2
#> 610        2      2
#> 611        2      4
#> 612        2      4
#> 613        2      4
#> 614        2      4
#> 615        2      4
#> 616        2      4
#> 617        2      3
#> 618        2      3
#> 619        3      3
#> 620        1      3
#> 621        2      3
#> 622        2      1
#> 623        2      3
#> 624        2      3
#> 625        3      4
#> 626        2      3
#> 627        2      4
#> 628        3      4
#> 629        3      2
#> 630       NA      3
#> 631       NA      3
#> 632        1      4
#> 633        1      2
#> 634        3      3
#> 635        3      4
#> 636        2      3
#> 637        2      3
#> 638        1      3
#> 639        2      2
#> 640        2      3
#> 641        2      3
#> 642        2      3
#> 643        2      3
#> 644        3      2
#> 645        2      2
#> 646        2      3
#> 647        2      2
#> 648        3      4
#> 649        2      2
#> 650        2      2
#> 651        2      2
#> 652        2      4
#> 653        2      4
#> 654        2      4
#> 655        2      1
#> 656        2      4
#> 657        3      3
#> 658        2      2
#> 659        1      3
#> 660        2      3
#> 661        2      2
#> 662        2      2
#> 663        3      2
#> 664        2      3
#> 665       NA      1
#> 666        3      3
#> 667        2      1
#> 668        1      2
#> 669        2      2
#> 670        2      3
#> 671        1      3
#> 672        2      2
#> 673        3      3
#> 674        1      4
#> 675        2      4
#> 676        1      1
#> 677        1      4
#> 678        2      4
#> 679        2      1
#> 680        1      3
#> 681        2      1
#> 682        2      2
#> 683        2      2
#> 684        2      4
#> 685        3      3
#> 686        2      1
#> 687        2      3
#> 688        2      3
#> 689        1      2
#> 690        2      4
#> 691        1      4
#> 692        3      3
#> 693        1      4
#> 694        2      3
#> 695        2      4
#> 696        2      4
#> 697        2      3
#> 698        3      4
#> 699        2      3
#> 700        2      4
#> 701        2      3
#> 702        2      2
#> 703        1      1
#> 704        2      2
#> 705       NA      2
#> 706        2      4
#> 707        1      2
#> 708        1      2
#> 709        1      3
#> 710        2      1
#> 711        1      2
#> 712        3      4
#> 713        2      3
#> 714        2      4
#> 715        2      4
#> 716        2      4
#> 717        2      4
#> 718        1      2
#> 719       NA      1
#> 720       NA      4
#> 721        2      2
#> 722        2      2
#> 723        1      3
#> 724        2      4
#> 725        2      3
#> 726        2      2
#> 727        2      4
#> 728        2      2
#> 729       NA      3
#> 730        2      3
#> 731        1      3
#> 732        1      4
#> 733        3      2
#> 734        2      4
#> 735        3      2
#> 736        2      3
#> 737       NA      2
#> 738        1      3
#> 739        3      2
#> 740        2      1
#> 741        2      4
#> 742        2      3
#> 743        2      3
#> 744        2      3
#> 745        1      4
#> 746        1      2
#> 747        2      2
#> 748        2      2
#> 749        3      3
#> 750        2      3
#> 751        1      4
#> 752        2      4
#> 753        3      4
#> 754        3      4
#> 755        1      4
#> 756        2      1
#> 757        2      2
#> 758        2      4
#> 759        2      4
#> 760        1      2
#> 761        2      2
#> 762        2      3
#> 763        2      2
#> 764        3      2
#> 765        2      2
#> 766        2      2
#> 767        1      3
#> 768        2      3
#> 769        3      1
#> 770        2      2
#> 771        2      1
#> 772        3      1
#> 773        2      2
#> 774        1      3
#> 775        1      2
#> 776        2      1
#> 777        1      3
#> 778       NA      3
#> 779        2      2
#> 780        2      2
#> 781        2      2
#> 782        2      2
#> 783       NA      3
#> 784        3      3
#> 785        2      3
#> 786        2      3
#> 787       NA      4
#> 788        3      2
#> 789        1      4
#> 790        2      1
#> 791        2      1
#> 792        1      1
#> 793        2      2
#> 794        2      4
#> 795        2      2
#> 796        2      2
#> 797        2      2
#> 798        2      2
#> 799        2      1
#> 800        2      3
#> 801        1      2
#> 802        1      4
#> 803        2      3
#> 804        1      4
#> 805        2      2
#> 806        2      3
#> 807        2      4
#> 808        2      3
#> 809        1      4
#> 810        2      3
#> 811        3      3
#> 812        2      3
#> 813        1      2
#> 814       NA      3
#> 815        1      2
#> 816        2      2
#> 817        1      3
#> 818        1      1
#> 819        2      1
#> 820        2      3
#> 821        2      1
#> 822        2      2
#> 823        2      2
#> 824        2      3
#> 825        2      2
#> 826        2      2
#> 827        2      2
#> 828        3      3
#> 829        3      2
#> 830        2      4
#> 831        3      1
#> 832        2      3
#> 833        2      1
#> 834        2      3
#> 835        2      1
#> 836        2      2
#> 837        2      1
#> 838        2      2
#> 839        2      2
#> 840        3      3
#> 841        2      4
#> 842        3      2
#> 843        2      2
#> 844        2      2
#> 845        2      1
#> 846        2      2
#> 847       NA      2
#> 848        3      2
#> 849        2      2
#> 850        3      2
#> 851        2      2
#> 852        3      4
#> 853        3      3
#> 854        2      1
#> 855        2      4
#> 856        3      2
#> 857        3      3
#> 858        3      3
#> 859        2      2
#> 860        3      2
#> 861        2      2
#> 862        2      3
#> 863        3      1
#> 864        2      2
#> 865        2      3
#> 866        2      3
#> 867        1      2
#> 868        1      2
#> 869        2      2
#> 870        2      1
#> 871        2      2
#> 872        2      2
#> 873        2      2
#> 874        3      3
#> 875        2      2
#> 876        1      2
#> 877        2      4
#> 878        2      3
#> 879        2      4
#> 880        2      2
#> 881        2      3
#> 882        2      1
#> 883        2      2
#> 884        1      3
#> 885        3      3
#> 886        3      2
#> 887        2      4
#> 888        2      3
#> 889        3      3
#> 890        3      2
#> 891        3      2
#> 892        2      3
#> 893        2      3
#> 894        3      3
#> 895        2      4
#> 896        2      2
#> 897        2      4
#> 898        1      3
#> 899        2      3
#> 900        1      4
#> 901        2      1
#> 902        2      2
#> 903       NA     NA
#> 904       NA     NA
#> 905       NA     NA
#> 906       NA     NA
#> 907       NA     NA
#> 908       NA     NA

# drop_labels() ----

if (require("sjmisc")) {
  rp <- rec_pattern(1, 100)
  rp

  # sample data
  data(efc)
  # recode carers age into groups of width 5
  x <- rec(efc$c160age, rec = rp$pattern)
  # add value labels to new vector
  x <- set_labels(x, labels = rp$labels)

  # watch result. due to recode-pattern, we have age groups with
  # no observations (zero-counts)
  frq(x)
  # now, let's drop zero's
  frq(drop_labels(x))
}
#> 
#> carer' age (x) <numeric>
#> # total N=908  valid N=901  mean=11.07  sd=2.69
#> 
#> Value | Label |   N | Raw % | Valid % | Cum. %
#> ----------------------------------------------
#>     4 | 16-20 |   5 |  0.55 |    0.55 |   0.55
#>     5 | 21-25 |  15 |  1.65 |    1.66 |   2.22
#>     6 | 26-30 |  28 |  3.08 |    3.11 |   5.33
#>     7 | 31-35 |  50 |  5.51 |    5.55 |  10.88
#>     8 | 36-40 |  51 |  5.62 |    5.66 |  16.54
#>     9 | 41-45 |  99 | 10.90 |   10.99 |  27.52
#>    10 | 46-50 | 126 | 13.88 |   13.98 |  41.51
#>    11 | 51-55 | 116 | 12.78 |   12.87 |  54.38
#>    12 | 56-60 | 122 | 13.44 |   13.54 |  67.92
#>    13 | 61-65 | 123 | 13.55 |   13.65 |  81.58
#>    14 | 66-70 |  82 |  9.03 |    9.10 |  90.68
#>    15 | 71-75 |  41 |  4.52 |    4.55 |  95.23
#>    16 | 76-80 |  32 |  3.52 |    3.55 |  98.78
#>    17 | 81-85 |   9 |  0.99 |    1.00 |  99.78
#>    18 | 86-90 |   2 |  0.22 |    0.22 | 100.00
#>  <NA> |  <NA> |   7 |  0.77 |    <NA> |   <NA>
#> 

# drop labels, also drop NA value labels, then also zap tagged NA
if (require("haven")) {
  x <- labelled(c(1:3, tagged_na("z"), 4:1),
                c("Agreement" = 1, "Disagreement" = 4, "Unused" = 5,
                  "Not home" = tagged_na("z")))
  x
  drop_labels(x, drop.na = FALSE)
  drop_labels(x)
  zap_na_tags(drop_labels(x))
}
#> <labelled<double>[8]>
#> [1]  1  2  3 NA  4  3  2  1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement

# fill_labels() ----

# create labelled integer, with tagged missings
if (require("haven")) {
  x <- labelled(
    c(1:3, tagged_na("a", "c", "z"), 4:1),
    c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
      "Refused" = tagged_na("a"), "Not home" = tagged_na("z"))
    )
  # get current values and labels
  x
  get_labels(x)

  fill_labels(x)
  get_labels(fill_labels(x))
  # same as
  get_labels(x, non.labelled = TRUE)
}
#> [1] "Agreement"    "2"            "3"            "Disagreement"