Drop, add or convert (non-)labelled values

For (partially) labelled vectors, zap_labels() will replace all values that have a value label attribute by NA; zap_unlabelled(), as counterpart, will replace all values that don't have a value label attribute by NA.

drop_labels() drops all value labels for unused values that have no cases (counts) in a vector. fill_labels() is the counterpart to drop_labels() and adds value labels to a partially labelled vector, i.e. if not all values are labelled, non-labelled values get labels.

drop_labels(x, ..., drop.na = TRUE)

fill_labels(x, ...)

zap_labels(x, ...)

zap_unlabelled(x, ...)

Arguments

x
(partially) labelled vector or a data frame with such vectors.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples'.

drop.na
Logical, whether existing value labels of tagged NA values (see tagged_na) should be removed (drop.na = TRUE, the default) or preserved (drop.na = FALSE). See get_na for more details on tagged NA values.

Value

For zap_labels(), x, where all labelled values are converted to NA.

For zap_unlabelled(), x, where all non-labelled values are converted to NA.

For drop_labels(), x, where value labels for non-existing values are removed.

For fill_labels(), x, where labels for non-labelled values are added.

If x is a data frame, the complete data frame x will be returned, with variables specified in ... being converted; if ... is not specified, applies to all variables in the data frame.

Examples

# zap_labels() ----

data(efc)
str(efc$e42dep)

#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:4] 1 2 3 4
#>   ..- attr(*, "names")= chr [1:4] "independent" "slightly dependent" "moderately dependent" "severely dependent"

x <- set_labels(
  efc$e42dep,
  labels = c("independent" = 1, "severe dependency" = 4)
)
table(x)

#> x
#>   1   2   3   4 
#>  66 225 306 304 

get_values(x)

#> [1] 1 4

str(x)

#>  num [1:908] 3 3 3 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# zap all labelled values
table(zap_labels(x))

#> 
#>   2   3 
#> 225 306 

get_values(zap_labels(x))

#> numeric(0)

str(zap_labels(x))

#>  num [1:908] 3 3 3 NA NA NA NA NA NA NA ...
#>  - attr(*, "labels")= Named num(0) 
#>   ..- attr(*, "names")= chr(0) 

# zap all unlabelled values
table(zap_unlabelled(x))

#> 
#>   1   4 
#>  66 304 

get_values(zap_unlabelled(x))

#> [1] 1 4

str(zap_unlabelled(x))

#>  num [1:908] NA NA NA 4 4 4 4 4 4 4 ...
#>  - attr(*, "label")= chr "elder's dependency"
#>  - attr(*, "labels")= Named num [1:2] 1 4
#>   ..- attr(*, "names")= chr [1:2] "independent" "severe dependency"

# in a pipe-workflow
library(dplyr)
efc %>%
  select(c172code, e42dep) %>%
  set_labels(
    e42dep,
    labels = c("independent" = 1, "severe dependency" = 4)
  ) %>%
  zap_labels()

#> # A tibble: 908 x 2
#>    c172code e42dep
#>       <dbl>  <dbl>
#>  1       NA      3
#>  2       NA      3
#>  3       NA      3
#>  4       NA     NA
#>  5       NA     NA
#>  6       NA     NA
#>  7       NA     NA
#>  8       NA     NA
#>  9       NA     NA
#> 10       NA     NA
#> # ... with 898 more rows

# drop_labels() ----

library(sjmisc)
rp <- rec_pattern(1, 100)
rp

#> $pattern
#> [1] "1:5=1;6:10=2;11:15=3;16:20=4;21:25=5;26:30=6;31:35=7;36:40=8;41:45=9;46:50=10;51:55=11;56:60=12;61:65=13;66:70=14;71:75=15;76:80=16;81:85=17;86:90=18;91:95=19;96:100=20;"
#> 
#> $labels
#>        1        2        3        4        5        6        7        8 
#>    "1-5"   "6-10"  "11-15"  "16-20"  "21-25"  "26-30"  "31-35"  "36-40" 
#>        9       10       11       12       13       14       15       16 
#>  "41-45"  "46-50"  "51-55"  "56-60"  "61-65"  "66-70"  "71-75"  "76-80" 
#>       17       18       19       20 
#>  "81-85"  "86-90"  "91-95" "96-100" 
#> 

# sample data
data(efc)
# recode carers age into groups of width 5
x <- rec(efc$c160age, rec = rp$pattern)
# add value labels to new vector
x <- set_labels(x, labels = rp$labels)

# watch result. due to recode-pattern, we have age groups with
# no observations (zero-counts)
frq(x)

#> 
#> # carer' age (x) <numeric> 
#> # total N=908  valid N=901  mean=11.07  sd=2.69
#>  
#>  val  label frq raw.prc valid.prc cum.prc
#>    1    1-5   0    0.00      0.00    0.00
#>    2   6-10   0    0.00      0.00    0.00
#>    3  11-15   0    0.00      0.00    0.00
#>    4  16-20   5    0.55      0.55    0.55
#>    5  21-25  15    1.65      1.66    2.22
#>    6  26-30  28    3.08      3.11    5.33
#>    7  31-35  50    5.51      5.55   10.88
#>    8  36-40  51    5.62      5.66   16.54
#>    9  41-45  99   10.90     10.99   27.52
#>   10  46-50 126   13.88     13.98   41.51
#>   11  51-55 116   12.78     12.87   54.38
#>   12  56-60 122   13.44     13.54   67.92
#>   13  61-65 123   13.55     13.65   81.58
#>   14  66-70  82    9.03      9.10   90.68
#>   15  71-75  41    4.52      4.55   95.23
#>   16  76-80  32    3.52      3.55   98.78
#>   17  81-85   9    0.99      1.00   99.78
#>   18  86-90   2    0.22      0.22  100.00
#>   19  91-95   0    0.00      0.00  100.00
#>   20 96-100   0    0.00      0.00  100.00
#>   NA     NA   7    0.77        NA      NA
#> 
#> 

# now, let's drop zero's
frq(drop_labels(x))

#> 
#> # carer' age (x) <numeric> 
#> # total N=908  valid N=901  mean=11.07  sd=2.69
#>  
#>  val label frq raw.prc valid.prc cum.prc
#>    4 16-20   5    0.55      0.55    0.55
#>    5 21-25  15    1.65      1.66    2.22
#>    6 26-30  28    3.08      3.11    5.33
#>    7 31-35  50    5.51      5.55   10.88
#>    8 36-40  51    5.62      5.66   16.54
#>    9 41-45  99   10.90     10.99   27.52
#>   10 46-50 126   13.88     13.98   41.51
#>   11 51-55 116   12.78     12.87   54.38
#>   12 56-60 122   13.44     13.54   67.92
#>   13 61-65 123   13.55     13.65   81.58
#>   14 66-70  82    9.03      9.10   90.68
#>   15 71-75  41    4.52      4.55   95.23
#>   16 76-80  32    3.52      3.55   98.78
#>   17 81-85   9    0.99      1.00   99.78
#>   18 86-90   2    0.22      0.22  100.00
#>   NA    NA   7    0.77        NA      NA
#> 
#> 

# drop labels, also drop NA value labels, then also zap tagged NA
library(haven)
x <- labelled(c(1:3, tagged_na("z"), 4:1),
              c("Agreement" = 1, "Disagreement" = 4, "Unused" = 5,
                "Not home" = tagged_na("z")))
x

#> <Labelled double>
#> [1]     1     2     3 NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>      5       Unused
#>  NA(z)     Not home

drop_labels(x, drop.na = FALSE)

#> <Labelled double>
#> [1]     1     2     3 NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>  NA(z)     Not home

drop_labels(x)

#> <Labelled double>
#> [1]     1     2     3 NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement

zap_na_tags(drop_labels(x))

#> <Labelled double>
#> [1]  1  2  3 NA  4  3  2  1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement

# fill_labels() ----

# create labelled integer, with tagged missings
library(haven)
x <- labelled(
  c(1:3, tagged_na("a", "c", "z"), 4:1),
  c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
    "Refused" = tagged_na("a"), "Not home" = tagged_na("z"))
  )
# get current values and labels
x

#> <Labelled double>
#>  [1]     1     2     3 NA(a) NA(c) NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      4 Disagreement
#>  NA(c)        First
#>  NA(a)      Refused
#>  NA(z)     Not home

get_labels(x)

#> [1] "Agreement"    "Disagreement"

fill_labels(x)

#> <Labelled double>
#>  [1]     1     2     3 NA(a) NA(c) NA(z)     4     3     2     1
#> 
#> Labels:
#>  value        label
#>      1    Agreement
#>      2            2
#>      3            3
#>      4 Disagreement
#>  NA(c)        First
#>  NA(a)      Refused
#>  NA(z)     Not home

get_labels(fill_labels(x))

#> [1] "Agreement"    "2"            "3"            "Disagreement"

# same as
get_labels(x, include.non.labelled = TRUE)

#> [1] "Agreement"    "2"            "3"            "Disagreement"